+++ sklearn/ensemble/__init__.py
@@ -18,8 +18,7 @@ from ._gb import GradientBoostingClassifier
 from ._gb import GradientBoostingRegressor
 from ._voting import VotingClassifier
 from ._voting import VotingRegressor
-from ._partial_dependence import partial_dependence
-from ._partial_dependence import plot_partial_dependence
+from . import partial_dependence
 
 __all__ = ["BaseEnsemble",
            "RandomForestClassifier", "RandomForestRegressor",
@@ -29,4 +28,4 @@ __all__ = ["BaseEnsemble",
            "GradientBoostingRegressor", "AdaBoostClassifier",
            "AdaBoostRegressor", "VotingClassifier", "VotingRegressor",
            "bagging", "forest", "gradient_boosting",
-           "partial_dependence", "plot_partial_dependence", "weight_boosting"]
+           "partial_dependence", "weight_boosting"]
